requirements.txt
python-telegram-bot==22.3
python-telegram-bot[job-queue]==22.3
requests
beautifulsoup4
schedule
certifi
watchdog
selenium
webdriver-manager